requirements.txt
# Binance API
python-binance==1.0.19

# Data processing
pandas==2.1.4
numpy==1.26.2

# Technical indicators
pandas-ta==0.3.14b0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging and monitoring
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

